requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
transformers==4.35.0
torch==2.4.1
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
reportlab==4.0.4
Pillow==10.0.1
python-dotenv==1.0.0
PyPDF2==3.0.1